feat: add SEO metadata to blog post pages
- Split blog/[slug]/page.tsx into server wrapper + client BlogPostContent
- Add generateMetadata with title, description, OpenGraph article tags
- Include publishedTime, authors, tags, featured image and canonical URL
- Revalidate metadata fetch hourly via next: { revalidate: 3600 }

closes #27
